check if zip lambda files exist before removing

+++ aimodelshare/api.py
@@ -188,9 +188,15 @@ def create_prediction_api(my_credentials, model_filepath, unique_model_id, model
     except Exception as e:
         print(e)
 
-    os.remove(os.path.join(temp_dir, 'archive.zip'))
-    os.remove(os.path.join(temp_dir,'main.py'))
-
+    import os
+    if os.path.exists(os.path.join(temp_dir, 'archive.zip')):
+      os.remove(os.path.join(temp_dir, 'archive.zip'))
+    else:
+      pass 	
+    if os.path.exists(os.path.join(temp_dir,'main.py')):
+      os.remove(os.path.join(temp_dir,'main.py'))
+    else:
+      pass         
 
 # Upload model eval lambda function zipfile to user's model file folder on s3
     if categorical == 'TRUE':
@@ -219,11 +225,31 @@ def create_prediction_api(my_credentials, model_filepath, unique_model_id, model
 
     except Exception as e:
         print(e)
+    
+    import os
+    if os.path.exists(os.path.join(temp_dir, 'archivetest.zip')):
+      os.remove(os.path.join(temp_dir, 'archivetest.zip'))
+    else:
+      pass 	
 
-    os.remove(os.path.join(temp_dir, 'archivetest.zip'))
-    os.remove(os.path.join(temp_dir, 'archive2.zip'))
-    os.remove(os.path.join(temp_dir, 'archive3.zip'))
-    os.remove(os.path.join(temp_dir,'main.py'))
+    if os.path.exists(os.path.join(temp_dir, 'archive2.zip')):
+      os.remove(os.path.join(temp_dir, 'archive2.zip'))
+    else:
+      pass         
+
+    if os.path.exists(os.path.join(temp_dir, 'archive2.zip')):
+      os.remove(os.path.join(temp_dir, 'archive2.zip'))
+    else:
+      pass 
+    if os.path.exists(os.path.join(temp_dir, 'archive3.zip')):
+      os.remove(os.path.join(temp_dir, 'archive3.zip'))
+    else:
+      pass 
+    if os.path.exists(os.path.join(temp_dir,'main.py')):
+      os.remove(os.path.join(temp_dir,'main.py'))
+    else:
+      pass 
+
 #!!! 2. update lambda creation code and iam policy attachements/apigateway integrations next.
 
     # Create and/or update roles for lambda function you will create below
